panther_point: Rework integrated graphics injection
Currently, integrated graphics injection in the DSDT checks for the
presence of a PCIe GPU. If one is found, a connectorless platform ID
is injected, which disables IGPU output while keeping QuickSync
available (for hardware-accelerated H.264 encoding, etc). However, this
prevents the use of both the IGPU and a PCIe GPU for display output,
and can cause panics if the user attempts to boot with a display
connected to the IGPU while a PCIe GPU is present (due to the
connectorless platform ID).

Rework the integrated graphics injection to always inject platform IDs
that support display output and remove the now-unnecessary checks for
PCIe GPUs.